moved generator templates and projects to project root
