+++ lib/cli/installMenu.js
@@ -1,11 +1,12 @@
 'use strict';
 
-var generatorUtil = require('../generator/generator').util,
-    inquirer      = require('inquirer'),
+var inquirer      = require('inquirer'),
     path          = require('path'),
-    fs            = require('fs'),
-    webStorm      = require('../generator/templates/webStorm'),
-    ideTemplate   = require('ide-template');
+    fs            = require('fs');
+
+var ideTemplate   = require('ide-template'),
+    generatorUtil = require('../generator/generator').util,
+    webStorm      = require('../../templates/webStorm');
 
 require('shelljs/global');
