+++ lib/generator/generator.js
@@ -35,20 +35,31 @@ var projects = [];
 }
 
 /**
- * The tools absolute template folder path for config files, ide settings etc...
- * @return {*|string}
+ * The absolute path for config files, ide settings etc...
+ * @param {...string} [suffix] Any number of path suffix elements
+ * @return {string}
  */
-function commonTemplatesPath() {
-  return path.join(__dirname, 'templates');
+function templatesPath(suffix) {
+  var elements = [__dirname, '..', '..', 'templates'].concat(Array.prototype.slice.call(arguments));
+  return path.resolve(path.join.apply(path, elements));
+}
+
+/**
+ * The absolute path for sample projects
+ * @param {...string} [suffix] Any number of path suffix elements
+ * @return {string}
+ */
+function projectsPath(suffix) {
+  var elements = [__dirname, '..', '..', 'projects'].concat(Array.prototype.slice.call(arguments));
+  return path.resolve(path.join.apply(path, elements));
 }
 
 /**
  * Using the defaultGlobalConfig() and template generate the global .angularity config file.
  * @param override template context
  */
- function createAngularityGlobalConfig(override) {
-  var configTemplatePath = path.join(commonTemplatesPath(), '.angularity');
-  var configTemplate = fs.readFileSync(configTemplatePath, 'utf8');
+function createAngularityGlobalConfig(override) {
+  var configTemplate = fs.readFileSync(templatesPath('.angularity'), 'utf8');
   var defaultContext = configDefaults.globalConfig;
 
   if (typeof override !== 'undefined') {
@@ -71,9 +82,8 @@ function commonTemplatesPath() {
  * @param destination
  * @param override
  */
- function createAngularityProjectConfig(destination, override) {
-  var configTemplatePath = path.join(commonTemplatesPath(), 'angularity.json');
-  var configTemplate = fs.readFileSync(configTemplatePath, 'utf8');
+function createAngularityProjectConfig(destination, override) {
+  var configTemplate = fs.readFileSync(templatesPath('angularity.json'), 'utf8');
   var defaultContext = configDefaults.projectConfig;
 
   if (typeof override !== 'undefined') {
@@ -107,10 +117,9 @@ function validateGeneratorProject(projectPath) {
  * @returns {Array}
  */
 function listProjects() {
-  var projectsPath = path.join(__dirname, 'projects');
-  return fs.readdirSync(projectsPath)
+  return fs.readdirSync(projectsPath())
     .filter(function (project) {
-      return validateGeneratorProject(path.join(projectsPath, project));
+      return validateGeneratorProject(projectsPath(project));
     });
 }
 
@@ -173,8 +182,7 @@ function validateExistingProject(directory) {
  */
 function requireProjects() {
   listProjects().forEach(function (project) {
-    var pathResolved = path.resolve(__dirname, '.' + path.sep, path.join('projects', project, 'index'));
-    require(pathResolved);
+    require(projectsPath(project, 'index'));
   });
 }
 
@@ -184,11 +192,6 @@ function requireProjects() {
  * @constructor
  */
 function GeneratorProject(type) {
-  var projectTypePath    = path.join(__dirname, 'projects', type);
-  var defaultProjectName = 'angularity-project';
-  if (!fs.existsSync(projectTypePath)) {
-    throw new Error('There does not seem to be a project with the name: ' + type);
-  }
 
   /**
    * The project type is the folder name of the generator project.
@@ -199,22 +202,25 @@ function GeneratorProject(type) {
   /**
    * The absolute path to the generator project template.
    */
-  this.projectTypePath = projectTypePath;
+  this.projectTypePath = projectsPath(type);
+  if (!fs.existsSync(this.projectTypePath)) {
+    throw new Error('There does not seem to be a project with the name: ' + type);
+  }
 
   /**
    * The project name used by the generator to populate the config files, destination etc.
    */
-  this.projectName = defaultProjectName;
+  this.projectName = 'angularity-project';
 
   /**
    * The absolute path the the generator project's template folder.
    */
-  this.templatePath = path.join(__dirname, 'projects', type, 'template');
+  this.templatePath = projectsPath(type, 'template');
 
   /**
    * The generator project's destination path.
    */
-  this.destination = path.join(String(process.cwd()), defaultProjectName);
+  this.destination = path.join(String(process.cwd()), this.projectName);
 
   // TODO block comment
   this.projectConfig = configDefaults.projectConfig;
@@ -259,7 +265,7 @@ module.exports = {
   requireProjects: requireProjects,
   util           : {
     generateProject              : generateProject,
-    commonTemplatesPath          : commonTemplatesPath,
+    templatesPath                : templatesPath,
     createAngularityGlobalConfig : createAngularityGlobalConfig,
     createAngularityProjectConfig: createAngularityProjectConfig,
     validateGeneratorProject     : validateGeneratorProject,
