+++ projects/minimal-es5/template/.jshintrc
