+++ projects/minimal-es5/template/bower.json
