+++ projects/minimal-es5/template/package.json
