+++ projects/minimal-es5/template/src/css-lib/site.scss
