+++ projects/minimal-es5/template/src/js-lib/app/appController.js
