+++ projects/minimal-es5/template/src/js-lib/app/appRoutes.js
