+++ projects/minimal-es6/index.js
@@ -7,14 +7,13 @@
 
 var gulp         = require('gulp'),
     fs           = require('fs'),
-    generator    = require('../../generator'),
-    config       = require('../../../config/config'),
-    configValues = require('../../../config/defaults'),
+    generator    = require('../../lib/generator/generator'),
+    config       = require('../../lib/config/config'),
+    configValues = require('../../lib/config/defaults'),
     template     = require('lodash.template'),
     webStorm     = require('ide-template').webStorm;
 
-gulp.task('minimal-es6', function ()
-{
+gulp.task('minimal-es6', function () {
   var project = generator.currentProject;
 
   project.copyProjectTemplateFiles();
