+++ projects/minimal-es6/projectPane.xml
