+++ projects/minimal-es6/template/.jshintrc
