+++ projects/minimal-es6/template/bower.json
