+++ projects/minimal-es6/template/package.json
