+++ projects/minimal-es6/template/src/js-lib/app/AppController.js
