+++ projects/minimal-es6/template/src/js-lib/utility/extend.js
