+++ templates/idea/fileTemplates/es5/es5-Factory.js
