+++ templates/idea/fileTemplates/es5/es5-TestSpec.js
