+++ templates/idea/fileTemplates/es6/es6-Controller.js.tpl
