+++ templates/idea/fileTemplates/es6/es6-Directive.js.tpl
