+++ templates/idea/fileTemplates/es6/es6-Factory.js.tpl
