+++ templates/idea/fileTemplates/es6/es6-testSpec.js
