+++ templates/idea/tools/Angularity.xml
